% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jackStraw2.R
\name{calc_PCpval}
\alias{calc_PCpval}
\title{Calculate p values for PCs given the rotated intact matrix and rotated
permuted matrix.}
\usage{
calc_PCpval(x, xp)
}
\arguments{
\item{x}{Subtset or entire \code{object@pca.x}}

\item{xp}{\code{fake.pcVals}}
}
\value{
p-value matrix with genes as rows and PCs on columns
}
\description{
Calculate p values for PCs given the rotated intact matrix and rotated
permuted matrix.
}

