% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eval_perf.R
\name{eval_roc_gene}
\alias{eval_roc_gene}
\title{Generate ROC for specific landmarkd gene}
\usage{
eval_roc_gene(object, g, g.staining)
}
\arguments{
\item{object}{seurat object}

\item{g}{gene name}

\item{g.staining}{Value of gene in in situ staining data}
}
\value{
roc object
}
\description{
Generate ROC for specific landmarkd gene
}

