% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eval_perf.R
\name{eval_seurat}
\alias{eval_seurat}
\title{Evaluate Seurat package performance}
\usage{
eval_seurat(object, genes.eval, dir, parallel = TRUE, scheme = c("lasso",
  "plsr"), insitu.path)
}
\arguments{
\item{object}{The seurat object.}

\item{genes.eval}{The landmark genes to be used to evaluted performance.}

\item{dir}{Directory to save the ROC curve plot and violin plot.}

\item{parallel}{If "TRUE", \link{foreach} is used to run parallel.}

\item{scheme}{Imputation strategies.}

\item{insitu.path}{Path to in situ staining data.}
}
\description{
Compare the inferred and observed in-situ patterns of landmark
genes to evaluate Seurat performance. Refer to paper 'Evaluating Seurat's
perforamce' for details.
}

