% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eval_perf.R
\name{eval_seurat_innter}
\alias{eval_seurat_innter}
\title{Run seurat analysis pipeline (optional: excluding specific landmark gene)}
\usage{
eval_seurat_innter(object, g = NULL, scheme = "lasso", insitu)
}
\arguments{
\item{object}{Object of seurat. Load the object before imputation.}

\item{g}{Specific landmark gene (upper case). If not setting, all landmark genes from in
situ data are used. Otherwise, \code{g} is excluded for generating spatial
reference map}

\item{scheme}{Imputation strategy. Default: lasso. Other options are: plsr,
tlasso}

\item{insitu}{Intact in situ staining data}
}
\value{
roc object
}
\description{
Run seurat analysis pipeline (optional: excluding specific landmark gene)
}

