% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/impute_expr.R
\name{fill_imputed_expr}
\alias{fill_imputed_expr}
\title{Calculate imputed expression values}
\usage{
fill_imputed_expr(object, genes.use = NULL, genes.fit = NULL,
  scheme = "lasso", s.use = 20, do.print = FALSE, gram = TRUE)
}
\arguments{
\item{object}{Seurat object}

\item{genes.use}{A vector of genes (predictors) that can be used for
building the LASSO models.}

\item{genes.fit}{A vector of genes to impute values for}

\item{scheme}{Imputation strategies: "lasso", "plsr".}

\item{s.use}{Maximum number of steps taken by the algorithm (lower values
indicate a greater degree of smoothing)}

\item{do.print}{Print progress (output the name of each gene after it has
been imputed).}

\item{gram}{The use.gram argument passed to lars}
}
\value{
Returns a Seurat object where the imputed values have been added to
object@data
}
\description{
Uses L1-constrained linear models (LASSO) to impute single cell
gene expression values.
}

