% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/map_cell_location.R
\name{initial_mapping}
\alias{initial_mapping}
\title{Infer spatial origins for single cells}
\usage{
initial_mapping(object, cells.use = NULL, gene.exclude = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{cells.use}{Which cells to map}
}
\value{
Seurat object, where mapping probabilities for each bin are stored
in object@final.prob
}
\description{
Probabilistically maps single cells based on (imputed) gene expression
estimates, a set of mixture models, and an in situ spatial reference map.
}

