% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/map_cell_location.R
\name{map_cell}
\alias{map_cell}
\title{Helper function of \code{initial_mapping}.}
\usage{
map_cell(object, cell.name, gene.exclude = NULL, do.plot = FALSE,
  safe.use = TRUE, text.val = NULL, do.rev = FALSE)
}
\description{
Map one cell to bins.
}

