% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/map_cell_location.R
\name{refined.mapping2}
\alias{refined.mapping2}
\title{Quantitative refinement of spatial inferences}
\usage{
refined.mapping2(object, genes.use)
}
\arguments{
\item{object}{Seurat object}

\item{genes.use}{Genes to use to drive the refinement procedure.}
}
\value{
Seurat object, where mapping probabilities for each bin are stored
in object@final.prob
}
\description{
Refines the initial mapping with more complex models that allow gene
expression to vary quantitatively across bins (instead of 'on' or 'off'),
and that also considers the covariance structure between genes.
}
\details{
Full details given in spatial mapping manuscript.
}

